Safari-611.1.10/LayoutTests/media/modern-media-controls/button/button-active-state-expected.txt

Testing an Button has a scale transform when pressed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(button.image.element).transform is "matrix(0.89, 0, 0, 0.89, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
